Simplified the implementation
