+++ android/app/src/main/AndroidManifest.xml
@@ -24,7 +24,6 @@
         </activity>
 
         <activity android:name=".component.ComponentActivity" />
-        <activity android:name=".component.NativeComponentActivity" />
     </application>
 
 </manifest>
