+++ android/app/src/main/java/com/sample/MainActivity.kt
@@ -2,32 +2,28 @@ package com.sample
 
 import android.os.Bundle
 import android.view.LayoutInflater
-import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.facebook.react.ReactActivity
 import com.sample.component.ComponentActivity
 import com.sample.component.ComponentListAdapter
 import com.sample.component.ComponentViewModel
-import com.sample.component.ComponentViewModel.NativeComponentViewModel
-import com.sample.component.ComponentViewModel.JsComponentViewModel
-import com.sample.component.NativeComponentActivity
 import com.sample.manifest.ManifestProvider
 import dagger.android.AndroidInjection
 import javax.inject.Inject
 
-class MainActivity : AppCompatActivity() {
+class MainActivity : ReactActivity() {
 
     @Inject
     lateinit var manifestProvider: ManifestProvider
 
     private val listener = { component: ComponentViewModel ->
-        val intent = when (component) {
-            is JsComponentViewModel -> ComponentActivity.newIntent(this, component.name, component.displayName)
-            is NativeComponentViewModel -> NativeComponentActivity.newIntent(this, component.displayName, component.klass)
-        }
-
-        startActivity(intent)
+        startActivity(
+            ComponentActivity.newIntent(
+                this, component.name, component.displayName
+            )
+        )
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -39,7 +35,7 @@ class MainActivity : AppCompatActivity() {
         val manifest = manifestProvider.manifest
             ?: throw IllegalStateException("app.json is not provided or TestApp is misconfigured")
         val components = manifest.components.map {
-            ComponentViewModel.Factory.create(it.key, it.value.displayName)
+            ComponentViewModel(it.key, it.value.displayName)
         }
 
         supportActionBar?.title = manifest.displayName
