+++ android/app/src/main/java/com/sample/component/ComponentActivity.kt
@@ -4,11 +4,12 @@ import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
 import android.view.MenuItem
+import androidx.fragment.app.Fragment
 import com.facebook.react.ReactActivity
-import com.facebook.react.modules.core.DefaultHardwareBackBtnHandler
 import com.facebook.soloader.SoLoader
 
-class ComponentActivity : ReactActivity(), DefaultHardwareBackBtnHandler {
+class ComponentActivity : ReactActivity() {
+
     companion object {
         private const val COMPONENT_NAME = "extra:componentName"
         private const val COMPONENT_DISPLAY_NAME = "extra:componentDisplayName"
@@ -29,14 +30,34 @@ class ComponentActivity : ReactActivity(), DefaultHardwareBackBtnHandler {
         super.onCreate(savedInstanceState)
         SoLoader.init(this, false)
 
-        title = intent.extras?.getString(COMPONENT_DISPLAY_NAME, null)
+        val componentDisplayName = intent.extras?.getString(COMPONENT_DISPLAY_NAME, null)
             ?: throw IllegalArgumentException("Component display name has to be provided.")
+        val componentName = intent.extras?.getString(COMPONENT_NAME, null)
+            ?: throw IllegalArgumentException("Class name has to be provided.")
+
+        title = componentDisplayName
 
         supportActionBar?.setHomeButtonEnabled(true)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
 
-        val componentName = intent.getStringExtra(COMPONENT_NAME)
-        loadApp(componentName)
+        findClass(componentName)?.let {
+            val fragmentClass = it as Class<Fragment>
+            val fragment = fragmentClass.newInstance()
+
+            if (savedInstanceState == null) {
+                supportFragmentManager.beginTransaction()
+                    .replace(android.R.id.content, fragment)
+                    .commitNow()
+            }
+        } ?: loadApp(componentName)
+    }
+
+    private fun findClass(name: String): Class<*>? {
+        return try {
+            Class.forName(name)
+        } catch (e: ClassNotFoundException) {
+            null
+        }
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
@@ -46,5 +67,13 @@ class ComponentActivity : ReactActivity(), DefaultHardwareBackBtnHandler {
         }
 
         return super.onOptionsItemSelected(item)
+    }
+
+    override fun getSystemService(name: String): Any? {
+        if (name == "service:reactNativeHostService") {
+            return reactNativeHost
+        }
+
+        return super.getSystemService(name)
     }
 }
