+++ android/app/src/main/java/com/sample/di/TestAppBindings.kt
@@ -4,7 +4,6 @@ import android.app.Application
 import android.content.Context
 import com.facebook.react.ReactNativeHost
 import com.sample.MainActivity
-import com.sample.component.NativeComponentActivity
 import com.sample.react.TestAppReactNativeHost
 import dagger.Binds
 import dagger.Module
@@ -22,8 +21,4 @@ abstract class TestAppBindings {
     @ActivityScope
     @ContributesAndroidInjector
     abstract fun contributeMainActivityInjector(): MainActivity
-
-    @ActivityScope
-    @ContributesAndroidInjector
-    abstract fun contributeNativeComponentActivityInjector(): NativeComponentActivity
 }
